SFX, Timeout, Background
I added an SFX audio bus for future options. I added a combo timeout that resets the combo after some time and also a visual indicator of hitting a combo and their combo expiring from them not pressing inputs for a while. I also used a shader to try to make a background. All in all feels pretty buggy, might need to go over it all a few more times to not be buggy, haha